restructure filesystem & tools as toolkit
