+++ packages/campfhir-cli/src/commands/assistant/tools/FhirQueryToolkit/index.ts
@@ -1,0 +1,36 @@
+import { JsonToolkit, Toolkit } from "langchain/agents";
+import { JsonObject, JsonSpec, Tool } from "langchain/tools";
+
+import { FhirAPI } from "./FhirAPI";
+import { FhirSummarizer } from "./FhirSummarizer";
+
+import { SessionLogger } from "../../helpers/sessionLogger";
+
+export class FhirQueryToolkit extends Toolkit {
+  tools: Tool[];
+
+  constructor() {
+    super();
+
+    const store = new JSONResponseStore();
+
+    this.tools = [new FhirAPI(store), new FhirSummarizer(store)];
+  }
+}
+
+export class JSONResponseStore {
+  jsonToolkit: JsonToolkit;
+  jsonSpec: JsonSpec;
+
+  constructor() {
+    this.jsonSpec = new JsonSpec({});
+    this.jsonToolkit = new JsonToolkit(this.jsonSpec);
+  }
+
+  setResponse(json: JsonObject) {
+    SessionLogger.separator();
+    SessionLogger.log("JSON: ", json);
+    SessionLogger.separator();
+    this.jsonSpec.obj = json;
+  }
+}
