--- packages/campfhir-cli/src/commands/assistant/tools/FhirURL.ts
@@ -1,328 +1,0 @@
-import * as fs from "fs";
-
-import { createClient, SupabaseClient } from "@supabase/supabase-js";
-import { AgentExecutor, ZeroShotAgent } from "langchain/agents";
-import { LLMChain } from "langchain/chains";
-import { Document } from "langchain/document";
-import { OpenAIEmbeddings } from "langchain/embeddings/openai";
-import { OpenAI } from "langchain/llms/openai";
-import { SupabaseHybridSearch } from "langchain/retrievers/supabase";
-import { type ChainValues } from "langchain/schema";
-import { Tool } from "langchain/tools";
-
-import dayjs from "dayjs";
-
-import { BufferMemory } from "langchain/memory";
-import { LoggingOutputParser } from "../parsers/LoggingOutputParser";
-import { fhirUrlAgentPrompt } from "../prompts/fhirUrlAgentPrompt";
-import { FhirAPI } from "./FhirAPI";
-import { FhirSummarizer, JSONResponseStore } from "./FhirSummarizer";
-
-export class FhirQuestion extends Tool {
-  name = "FhirQuestion";
-  description =
-    "Useful for answering questions about medical data stored on a FHIR RESTful API server.  The input to this tool should be a natural language query about some FHIR resource.  The output of this tool is the summarized Fhir RESTFul API server response.";
-
-  tools: Tool[];
-  executor: AgentExecutor | undefined;
-
-  constructor() {
-    super();
-
-    const store = new JSONResponseStore();
-
-    this.tools = [
-      new KnownEndpoints(),
-      new EndpointParams(),
-      new EndpointParameterDetails(),
-      new DateFormat(),
-      new FhirAPI(store),
-      new FhirSummarizer(store),
-    ];
-  }
-
-  async _call(input: string): Promise<string> {
-    if (!this.executor) {
-      this.executor = await this.initializeAgent();
-    }
-
-    try {
-      const response: ChainValues = await this.executor.call({
-        input,
-      });
-
-      return response.output;
-    } catch (error) {
-      console.log("error: ", error);
-      return `Sorry I don't know about: ${input}`;
-    }
-  }
-
-  protected async initializeAgent() {
-    const llm = new OpenAI({ temperature: 0 });
-    const prompt = await fhirUrlAgentPrompt(this.tools);
-    const llmChain = new LLMChain({
-      llm,
-      prompt,
-      // verbose: true,
-    });
-    const agent = new ZeroShotAgent({
-      llmChain,
-      allowedTools: this.tools.map((tool) => tool.name),
-      outputParser: new LoggingOutputParser("FhirQuestion"),
-    });
-    const memory = new BufferMemory({ memoryKey: "chat_history" });
-    return AgentExecutor.fromAgentAndTools({
-      agent,
-      tools: this.tools,
-      memory,
-      //returnIntermediateSteps: true,
-    });
-  }
-}
-
-function searchParamsFromResource(resource: {
-  id: string;
-  code: string;
-  type: string;
-  expression: string;
-  description: string;
-  xpath: string;
-  xpathUsage: string;
-  base: string[];
-  comparator: string[];
-  target: string[];
-}) {
-  const {
-    id,
-    code,
-    type,
-    expression,
-    description,
-    xpath,
-    xpathUsage,
-    base,
-    comparator,
-    target,
-  } = resource;
-
-  const searchParams = {
-    id,
-    code,
-    type,
-    expression,
-    description,
-    xpath,
-    xpathUsage,
-    base,
-    comparator: comparator || undefined,
-    target: target || undefined,
-  };
-
-  return searchParams;
-}
-
-const searchParamsWhitelist = [
-  "_id",
-  "_summary",
-  "_type",
-  "account",
-  "active",
-  "actor",
-  "address",
-  "appointment",
-  "birthdate",
-  "category",
-  "class",
-  "condition",
-  "date",
-  "diagnosis",
-  "email",
-  "encounter",
-  "family",
-  "gender",
-  "given",
-  "goal",
-  "intent",
-  "length",
-  "location-period",
-  "location",
-  "name",
-  "participant-type",
-  "participant",
-  "patient",
-  "performer",
-  "phone",
-  "practitioner",
-  "reason-code",
-  "service-provider",
-  "status",
-  "subject",
-  "title",
-  "topic",
-  "type",
-  "url",
-  "value",
-]; // TODO: add more
-// filtered to reduce the size of the tools responses
-function filterParamByWhitelist(entries: any[]) {
-  return entries.filter((entry: any) =>
-    searchParamsWhitelist.includes(entry.resource.code)
-  );
-}
-
-let paramsByBase: any;
-function compileSearchParamsByBase(searchParamsSpec: any) {
-  if (!paramsByBase) {
-    paramsByBase = {};
-    const resourceParams = filterParamByWhitelist(searchParamsSpec.entry)
-      .filter((entry: any) => entry.resource.base.includes("Resource"))
-      .reduce((acc: any, entry: any) => {
-        const resource = entry.resource;
-        const searchParams = searchParamsFromResource(resource);
-        acc[searchParams.code] = searchParams;
-        return acc;
-      }, {});
-
-    filterParamByWhitelist(searchParamsSpec.entry).forEach((entry: any) => {
-      const resource = entry.resource;
-      resource.base.forEach((resourceType: string) => {
-        if (!paramsByBase[resourceType]) {
-          paramsByBase[resourceType] = resourceParams;
-        }
-
-        const searchParams = searchParamsFromResource(resource);
-
-        paramsByBase[resourceType][searchParams.code] = searchParams;
-      });
-    });
-  }
-
-  return paramsByBase;
-}
-
-class KnownEndpoints extends Tool {
-  name = "KnownEndpoints";
-  description =
-    "Useful for finding the known FHIR ENDPOINTS.  The input to this tool should be a natural language query about some FHIR resource.  The output of this tool is a deterministic JSON list of known FHIR ENDPOINTS.";
-
-  supabase: SupabaseClient;
-  retriever: SupabaseHybridSearch | undefined;
-
-  constructor() {
-    super();
-
-    this.supabase = createClient(
-      process.env.PUBLIC_SUPABASE_URL || "",
-      process.env.PUBLIC_SUPABASE_ANON_KEY || ""
-    );
-  }
-
-  async _call(input: string): Promise<string> {
-    const endpoints = await this.retrieveKnowEndpointsFor(input);
-    return JSON.stringify(endpoints);
-  }
-
-  protected async retrieveKnowEndpointsFor(input: string): Promise<string[]> {
-    if (!this.retriever) {
-      this.retriever = await this.newRetriever(this.supabase);
-    }
-
-    const response = await this.retriever.getRelevantDocuments(input);
-
-    const endpoints = [
-      ...new Set(
-        response
-          .map((doc: Document) => {
-            const { base, target } = doc.metadata;
-            return [base as string, target as string];
-          })
-          .flat()
-          .filter((x: any) => x)
-      ),
-    ];
-    return endpoints;
-  }
-
-  protected async newRetriever(client: SupabaseClient) {
-    return new SupabaseHybridSearch(new OpenAIEmbeddings(), {
-      client,
-      tableName: "documents",
-      similarityQueryName: "match_documents",
-      keywordQueryName: "kw_match_documents",
-      similarityK: 10,
-      keywordK: 10,
-    });
-  }
-}
-
-class EndpointParams extends Tool {
-  name = "EndpointParams";
-  description =
-    "Useful for finding the valid parameters for a given FHIR ENDPOINT.  The input to this tool should be a FHIR ENDPOINT name.  The output of this tool is a deterministic JSON list of valid parameters for the given FHIR ENDPOINT.";
-
-  searchParamsByBase: any;
-
-  constructor() {
-    super();
-    const jsonText = fs.readFileSync(
-      "/workspace/data/fhir-r4b-search-parameters.json",
-      "utf8"
-    );
-    const searchParamsSpec = JSON.parse(jsonText);
-
-    this.searchParamsByBase = compileSearchParamsByBase(searchParamsSpec);
-  }
-
-  async _call(input: string): Promise<string> {
-    return JSON.stringify(this.paramsForBase(input));
-  }
-
-  protected paramsForBase(base: string): string[][] {
-    const baseParams = Object.values(this.searchParamsByBase[base] || {});
-    return baseParams.map((param: any) => {
-      return param.code;
-      // const { code, description } = param;
-      // return [base, code, description];
-    });
-  }
-}
-
-class EndpointParameterDetails extends Tool {
-  name = "EndpointParameterDetails";
-  description =
-    "Useful for finding the details of a given FHIR PARAMETER.  The input to this tool should be a FHIR ENDPOINT & PARAMETER name pair, joined by a column (:).  The input format is ENDPOINT:PARAMETER.  The output of this tool is the deterministic usage properties of the given FHIR PARAMETER, as JSON.";
-
-  searchParamsByBase: any;
-
-  constructor() {
-    super();
-    const jsonText = fs.readFileSync(
-      "/workspace/data/fhir-r4b-search-parameters.json",
-      "utf8"
-    );
-    const searchParamsSpec = JSON.parse(jsonText);
-
-    this.searchParamsByBase = compileSearchParamsByBase(searchParamsSpec);
-  }
-
-  async _call(input: string): Promise<string> {
-    if (!input.includes(":")) {
-      return "Error: input must be of the form <base>:<param>";
-    }
-
-    const [base, param] = input.split(":");
-    const paramSpec = this.searchParamsByBase[base as string][param as string];
-    return JSON.stringify(paramSpec);
-  }
-}
-
-class DateFormat extends Tool {
-  name = "DateFormat";
-  description =
-    "Useful for formatting a date.  The input to this tool should be a date string.  The output of this tool is a date in the format YYYY-MM-DD.";
-
-  async _call(input: string): Promise<string> {
-    return dayjs(input).format("YYYY-MM-DD");
-  }
-}
